drop Oracle\AbstractQuery class
